Make doubleNAN an actual double, duh. (what one finds in reading the context lines in a checkin diff)

+++ src/nidas/core/Sample.cc
@@ -19,7 +19,7 @@
 
 const float nidas::core::floatNAN = nanf("");
 
-const float nidas::core::doubleNAN = nan("");
+const double nidas::core::doubleNAN = nan("");
 
 using namespace nidas::core;
 using namespace std;
